Fix for Clever section teachers
Some systems only support a single teacher, and they pull the first teacher in the list. Some of our classes need to have multiple teachers, even if those teachers don't even work there anymore, and since they are still in the mix, third parties that associate the class to the first teacher in the list are pulling the wrong staff member.
This change will always list the "default staff" for the class _first_, and will ignore classes that have no default class set. This isn't ideal, but it's the best we can do with SIS right now.